ApplicationListView and ApplicationView are now in Vue files
