+++ jstests/tests/user/test_application_list_view.js
@@ -1,6 +1,6 @@
 var Vue = require("vuejs");
 var models = require("user/models");
-var applicationListView = require("user/views/application_list_view");
+var ApplicationListView = require("user/views/application_list_view");
 require("filters");
 
 QUnit.module("user.app_list_view");
@@ -8,7 +8,7 @@ QUnit.test("rendering list", function (assert) {
     var done = assert.async();
 
     var model = new models.ApplicationListModel();
-    var appListView = new applicationListView.ApplicationListView({
+    var appListView = new ApplicationListView({
         data: function() { return { model: model }; }
     }).$mount();
 
@@ -40,7 +40,7 @@ QUnit.test("rendering nothing in the list", function (assert) {
     var done = assert.async();
 
     var model = new models.ApplicationListModel();
-    var appListView = new applicationListView.ApplicationListView({
+    var appListView = new ApplicationListView({
         data: function() { return { model: model }; }
     }).$mount();
 
@@ -68,7 +68,7 @@ QUnit.test("search form", function (assert) {
     var done = assert.async();
 
     var model = new models.ApplicationListModel();
-    var appListView = new applicationListView.ApplicationListView({
+    var appListView = new ApplicationListView({
         data: function() { return { model: model }; }
     }).$mount();
